# src/integration-test/resources/openapi.yml
openapi: 3.0.1
info:
  title: api-service-messaging-settings
  contact: {}
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: "2.0"
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /{municipalityId}:
    get:
      tags:
      - Messaging Settings
      summary: Get messaging settings
      description: Get all messaging settings or the ones that matches provided filter
      operationId: fetchMessagingSettings
      parameters:
      - name: municipalityId
        in: path
        description: Municipality ID
        required: true
        schema:
          type: string
        example: 2281
      - name: filter
        in: query
        description: "Syntax description: [spring-filter](https://github.com/turkraft/spring-filter/blob/85730f950a5f8623159cc0eb4d737555f9382bb7/README.md#syntax)"
        required: false
        schema:
          type: string
        example: "created > '2022-09-08T12:00:00.000+02:00' and values.key: 'namespace'\
          \ and values.value: 'NS1'"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MessagingSettings"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
  /{municipalityId}/user:
    get:
      tags:
      - Messaging Settings
      summary: Get messaging settings for organization connected to a user
      description: Get messaging settings for the organization connected to the provided
        user.
      operationId: getMessagingSettingsForUser
      parameters:
      - name: X-Sent-By
        in: header
        description: User identity
        required: true
        schema:
          type: string
        example: joe01doe;type=adAccount
      - name: municipalityId
        in: path
        description: Municipality ID
        required: true
        schema:
          type: string
        example: 2281
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MessagingSettings"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
components:
  schemas:
    Problem:
      type: object
      properties:
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties:
            type: object
        status:
          $ref: "#/components/schemas/StatusType"
        title:
          type: string
        detail:
          type: string
    StatusType:
      type: object
      properties:
        reasonPhrase:
          type: string
        statusCode:
          type: integer
          format: int32
    ConstraintViolationProblem:
      type: object
      properties:
        cause:
          $ref: "#/components/schemas/ThrowableProblem"
        stackTrace:
          type: array
          items:
            type: object
            properties:
              classLoaderName:
                type: string
              moduleName:
                type: string
              moduleVersion:
                type: string
              methodName:
                type: string
              fileName:
                type: string
              lineNumber:
                type: integer
                format: int32
              className:
                type: string
              nativeMethod:
                type: boolean
        type:
          type: string
          format: uri
        status:
          $ref: "#/components/schemas/StatusType"
        violations:
          type: array
          items:
            $ref: "#/components/schemas/Violation"
        title:
          type: string
        message:
          type: string
        instance:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties:
            type: object
        detail:
          type: string
        suppressed:
          type: array
          items:
            type: object
            properties:
              stackTrace:
                type: array
                items:
                  type: object
                  properties:
                    classLoaderName:
                      type: string
                    moduleName:
                      type: string
                    moduleVersion:
                      type: string
                    methodName:
                      type: string
                    fileName:
                      type: string
                    lineNumber:
                      type: integer
                      format: int32
                    className:
                      type: string
                    nativeMethod:
                      type: boolean
              message:
                type: string
              localizedMessage:
                type: string
        localizedMessage:
          type: string
    ThrowableProblem:
      type: object
      properties:
        cause:
          $ref: "#/components/schemas/ThrowableProblem"
        stackTrace:
          type: array
          items:
            type: object
            properties:
              classLoaderName:
                type: string
              moduleName:
                type: string
              moduleVersion:
                type: string
              methodName:
                type: string
              fileName:
                type: string
              lineNumber:
                type: integer
                format: int32
              className:
                type: string
              nativeMethod:
                type: boolean
        message:
          type: string
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties:
            type: object
        status:
          $ref: "#/components/schemas/StatusType"
        title:
          type: string
        detail:
          type: string
        suppressed:
          type: array
          items:
            type: object
            properties:
              stackTrace:
                type: array
                items:
                  type: object
                  properties:
                    classLoaderName:
                      type: string
                    moduleName:
                      type: string
                    moduleVersion:
                      type: string
                    methodName:
                      type: string
                    fileName:
                      type: string
                    lineNumber:
                      type: integer
                      format: int32
                    className:
                      type: string
                    nativeMethod:
                      type: boolean
              message:
                type: string
              localizedMessage:
                type: string
        localizedMessage:
          type: string
    Violation:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
    MessagingSettingValue:
      type: object
      properties:
        key:
          type: string
          description: Identifier key for the value setting
          readOnly: true
          example: department_name
        value:
          type: string
          description: Stored value for the value setting
          readOnly: true
          example: Department 44
        type:
          type: string
          description: "Type of data for the stored value. Can be one of [STRING|NUMERIC|BOOLEAN]"
          readOnly: true
          example: STRING
      description: Messaging setting value model
    MessagingSettings:
      type: object
      properties:
        id:
          type: string
          description: Id for the messaging setting instance
          readOnly: true
          example: c9383d10-6fb5-4fc1-bd0a-50bf5a24d5b7
        municipalityId:
          type: string
          description: Municipality id for municipality that the messaging setting
            instance belongs to
          readOnly: true
          example: "2281"
        created:
          type: string
          description: Timestamp when the instance was created
          format: date-time
          readOnly: true
          example: 2025-10-24T15:30:00+02:00
        updated:
          type: string
          description: Timestamp when the instance was last updated. Null if instance
            has never been updated.
          format: date-time
          readOnly: true
          example: 2025-10-25T16:30:00+02:00
        values:
          type: array
          description: Values connected to the messaging setting instance
          readOnly: true
          items:
            $ref: "#/components/schemas/MessagingSettingValue"
      description: Messaging settings response model
...
  securitySchemes: {}